feat: 세션방식 소셜로그인으로 변경
1. db 어댑터를 이용한 세션방식 로그인으로 변경
2. 회원가입시 db에 회원정보 및 계정정보, 세션정보가 생성
3. 해당정보를 기준으로 사용자 검증

+++ forum/pages/api/post/delete.js
@@ -1,7 +1,12 @@
 import { connectDB } from '@/util/database';
 import { ObjectId } from 'mongodb';
+import { getServerSession } from 'next-auth';
+import { authOptions } from '../auth/[...nextauth]';
 
 export default async function handler(요청, 응답) {
+    let session = await getServerSession(요청, 응답, authOptions);
+    console.log(session);
+
     if (요청.method === 'DELETE') {
         console.log(JSON.parse(요청.body));
 
@@ -11,7 +16,7 @@ export default async function handler(요청, 응답) {
         try {
             const client = await connectDB;
             const db = client.db('forum');
-            let result = await db.collection('post').deleteOne({ _id: new ObjectId(_id) });
+            let result = await db.collection('post').deleteOne({ _id: new ObjectId(_id), author: session.user.email });
             // console.log(result);
             if (!result.deletedCount) {
                 return 응답.status(500).json('삭제 실패');
